requirements.txt
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
cssselect==1.2.0
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
lxml==5.2.2
msal==1.30.0
nested-lookup==0.2.25
oauthlib==3.2.2
packaging==24.1
parsel==1.9.1
pip==24.0
playwright==1.45.1
proto-plus==1.24.0
protobuf==5.27.2
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyee==11.1.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
telegram==0.0.1
Telethon==1.36.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
w3lib==2.2.1
wheel==0.43.0
emoji==1.4.1
